refactor: 클래스 디렉터리 변경 HangmanGameRoundTable, HangmanGameTable -> hangman에서 domain으로 이동 RandomWordChooser -> hangman에서 util로 이동
